rename underscore for cli
