+++ limix/_cli/__init__.py
@@ -1,3 +1,3 @@
-from ._cli import cli
+from .cli import cli
 
 __all__ = ["cli"]
