--- limix/_cli/_scan/__init__.py
@@ -1,3 +1,0 @@
-from ._scan import scan
-
-__all__ = ["scan"]
