+++ limix/_cli/test/test_pipeline.py
@@ -1,0 +1,4 @@
+from limix._cli.pipeline import Pipeline
+
+def test_pipeline():
+    pass
