+++ limix/io/__init__.py
@@ -1,9 +1,28 @@
-from __future__ import absolute_import
+"""
+io module
+=========
 
-from . import bgen, bimbam, csv, hdf5, npy, plink
-from . import gen
-from ._fetch import fetch
+Subpackage for I/O operations.
+
+Modules
+-------
+bgen
+csv
+gen
+hdf5
+npy
+plink
+bimbam
+
+Functions
+---------
+fetch
+infer_filetype
+"""
+
+from . import bgen, bimbam, csv, gen, hdf5, npy, plink
 from ._detect import infer_filetype
+from ._fetch import fetch
 
 __all__ = [
     "bgen",
